requirements.txt
openai
python-dotenv
python-docx
PyPDF2
agno[azure]
streamlit
